Restored old update behavior
I.e. model increments state counter after each decoding, not only when the
update is nonempty.

Intuitively, this should be good when the decoder found a correct parse
because it gives a higher weight in the averaged model to the weights used
for this correct parse. When no contrasting pair of parses is found, it might
be better to not increment the state counter - remains to be investigated
some time maybe.